Changed namespace and, as class name `Calculations` was the same as the namespace changed it to `Calculator`
